Renamed the namespace for LabelDtos
