+++ Todo.Core/Dtos/LabelDtos/CreateLabelDto.cs
@@ -1,14 +1,13 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Todo.Core.Dtos.LabelDto;
+namespace Todo.Core.Dtos.LabelDtos;
 
-public class UpdateLabelDto
+public class CreateLabelDto
 {
-    public string? Name { get; set; } = null;
+    public string Name { get; set; } = string.Empty;
 
     [StringLength(7)]
     [RegularExpression("^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
         ErrorMessage = "Color must be a valid hex color code. Did you remember to include the '#'?")]
-
     public string? Color { get; set; } = null;
 }
